requirements.txt
dnspython==2.4.2
ns1_python==0.18.0
Requests==2.31.0
